Solved Login Error and Upload Images from Device Error
